refactor roll-out feature to be endpoint
