+++ next.config.js
@@ -5,9 +5,6 @@ const config = {
   images: {
     domains: ['i.ibb.co', 'cdn.sanity.io'],
   },
-  experimental: {
-    serverActions: true,
-  },
 };
 
 module.exports = withLinaria(config);
